move shared/CloseButton to shared/Button/Close
