+++ src/components/Editor/SearchBar.js
@@ -5,7 +5,7 @@ const Svg = require("../utils/Svg");
 const { find, findNext, findPrev, removeOverlay } = require("../../utils/source-search");
 const classnames = require("classnames");
 const { debounce, escapeRegExp } = require("lodash");
-const CloseButton = require("../shared/CloseButton");
+const CloseButton = require("../shared/Button/Close");
 const { isEnabled } = require("devtools-config");
 const ImPropTypes = require("react-immutable-proptypes");
