+++ src/components/Expressions.js
@@ -4,7 +4,7 @@ const { bindActionCreators } = require("redux");
 const ImPropTypes = require("react-immutable-proptypes");
 const actions = require("../actions");
 const { getExpressions, getLoadedObjects, getPause } = require("../selectors");
-const CloseButton = React.createFactory(require("./shared/CloseButton"));
+const CloseButton = React.createFactory(require("./shared/Button/Close"));
 const ObjectInspector = React.createFactory(require("./shared/ObjectInspector"));
 const { DOM: dom, PropTypes } = React;
