+++ src/components/shared/Button/Close.js
@@ -1,8 +1,8 @@
 const React = require("react");
 const { DOM: dom, PropTypes } = React;
-const Svg = require("../utils/Svg");
+const Svg = require("../../utils/Svg");
 
-require("./CloseButton.css");
+require("./Close.css");
 
 function CloseButton({ handleClick, buttonClass, tooltip }) {
   return dom.div({
